icetime: fixed chipdb path issue for Arch Linux
Icetime was hardcoding the path to chipdb as /usr/local/share causing
icetime to not find the chip db on Arch Linux where the prefix is just
/usr/share.

With this commit the PREFIX is passed as a preprocessor define and used
in icetime.cc to create the correct path. I don't know what the
canonical way of dealing with this is, but this was the least intrusive
way I could think of to get this fixed.